Add a resolver to fix variable binding
This implements the resolver class as described in
https://craftinginterpreters.com/resolving-and-binding.html

Also, it fixes a small previous inconsistency with the book.
I was passing in the string for a variable name in places
where the book used the token object. This now started to
cause more divergence from the book in my code and added
small complexity in some places. So I've refactored it to
match the book and use the token object instead of a string
for variable name.