Finish setting core files to strict typing
The only core files which are still not strict
are statements.rb and expressions.rb and
the problem with them is that they do some
(light) metaprogramming and sorbet has a very
difficult time with metaprogramming. To make it
work I'd need to define RBI files for them where
all methods on all classes are essentially "unrolled"
as if I typed them out manually.

If I'm doing that I'd want to generate it because
there's hardly anything in the actual files, both
methods are one-liners so if I was to generate RBI
files automatically with a script I could then make
it actually simpler by generating the ruby file
itself with all classes defined without metaprogramming
and I decided to stop short of that.