Raise branch coverage to maximum
For this commit I've enabled branch coverage on simplecov
and raise to maximu. 2 branches I couldn't cover because
they are both case statements and branch coverage is
saying we are not triggering the else branch. However, both
cases are exhaustive, covering all of the possible values
that can reach it. I would have to introduce a bug in the
rest of the code to be able to write a test to reach the
else branch at that point.

Overall, this branch of the experiment found the most issues,
more than type checkers.